daegon take

+++ src/main/java/pet/controller/PetDelete.java
@@ -40,9 +40,11 @@ public class PetDelete extends HttpServlet {
 		
 		int result = new PetService().petDelete(pet);
 		if (result > 0) {
+			if(!request.getParameter("deleteImg").equals("882.jpg")) {
 			String renameFileName = request.getParameter("deleteImg");
 			String savePath = request.getSession().getServletContext().getRealPath("/resources/images/petImg");
 			new File(savePath + "\\" + renameFileName).delete(); 
+			}
 			String link = "/petmily/mypet?memberSeq="+pet.getMemberSeq();
 			response.sendRedirect(link);
 		} else {
